Remove some code duplication
