+++ src/routes/join/nonprofits/+page.server.ts
@@ -1,8 +1,7 @@
-import { contentWrapper } from "$lib/content/contentful";
-import type {
-  NonprofitTestimonialProject,
-  Project,
-} from "src/lib/utils/schema";
+import {
+  contentWrapper,
+  getTestimonialNonprofit,
+} from "$lib/content/contentful";
 import type { PageServerLoad } from "./$types";
 
 export const load: PageServerLoad = () => {
@@ -18,19 +17,3 @@ export const load: PageServerLoad = () => {
     testimonialNonprofit: getTestimonialNonprofit(),
   };
 };
-
-async function getTestimonialNonprofit(): Promise<NonprofitTestimonialProject> {
-  const projects: Project[] = await contentWrapper.get("project");
-  const testimonialProjects = projects.filter(
-    (project) =>
-      project.testimonial !== undefined &&
-      project.testimonialSourceName !== undefined &&
-      project.testimonialSourceDescription !== undefined &&
-      project.testimonialSourceType === "Nonprofit"
-  ) as NonprofitTestimonialProject[];
-
-  const featuredTestimonialProject =
-    testimonialProjects[Math.floor(Math.random() * testimonialProjects.length)];
-
-  return featuredTestimonialProject;
-}
