+++ src/routes/join/sponsors/+page.server.ts
@@ -1,5 +1,7 @@
-import { contentWrapper } from "$lib/content/contentful";
-import type { NonprofitTestimonialProject, Project } from "$lib/utils/schema";
+import {
+  contentWrapper,
+  getTestimonialNonprofit,
+} from "$lib/content/contentful";
 
 import type { PageServerLoad } from "./$types";
 
@@ -13,19 +15,3 @@ export const load: PageServerLoad = () => {
     testimonialNonprofit: getTestimonialNonprofit(),
   };
 };
-
-async function getTestimonialNonprofit(): Promise<NonprofitTestimonialProject> {
-  const projects: Project[] = await contentWrapper.get("project");
-  const testimonialProjects = projects.filter(
-    (project) =>
-      project.testimonial !== undefined &&
-      project.testimonialSourceName !== undefined &&
-      project.testimonialSourceDescription !== undefined &&
-      project.testimonialSourceType === "Nonprofit"
-  ) as NonprofitTestimonialProject[];
-
-  const featuredTestimonialProject =
-    testimonialProjects[Math.floor(Math.random() * testimonialProjects.length)];
-
-  return featuredTestimonialProject;
-}
